# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Monocle",
    "description": "Personal knowledge capture and retrieval API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/notes": {
      "get": {
        "tags": [
          "notes"
        ],
        "summary": "List Notes",
        "operationId": "list_notes_api_notes_get",
        "parameters": [
          {
            "name": "folder",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Folder"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "updated",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_NoteRef_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/notes/{path}/backlinks": {
      "get": {
        "tags": [
          "notes"
        ],
        "summary": "Get Note Backlinks",
        "description": "Return all notes that link to `path` via structured links, wikilinks, or people co-mention.",
        "operationId": "get_note_backlinks_api_notes__path__backlinks_get",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BacklinkRef"
                  },
                  "title": "Response Get Note Backlinks Api Notes  Path  Backlinks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/notes/{path}": {
      "get": {
        "tags": [
          "notes"
        ],
        "summary": "Get Note",
        "operationId": "get_note_api_notes__path__get",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "notes"
        ],
        "summary": "Put Note",
        "operationId": "put_note_api_notes__path__put",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "notes"
        ],
        "summary": "Patch Note",
        "operationId": "patch_note_api_notes__path__patch",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotePatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "notes"
        ],
        "summary": "Delete Note",
        "operationId": "delete_note_api_notes__path__delete",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/notes/{path}/move": {
      "post": {
        "tags": [
          "notes"
        ],
        "summary": "Move Note",
        "operationId": "move_note_api_notes__path__move_post",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteMoveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Move Note Api Notes  Path  Move Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BacklinkRef": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "relation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relation"
          },
          "context": {
            "type": "string",
            "title": "Context",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "source"
        ],
        "title": "BacklinkRef"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LinkRef": {
        "properties": {
          "target": {
            "type": "string",
            "title": "Target"
          },
          "relation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relation"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "target"
        ],
        "title": "LinkRef",
        "description": "A structured link from a note's `links:` frontmatter field."
      },
      "Note": {
        "properties": {
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "metadata": {
            "$ref": "#/components/schemas/NoteMetadata"
          },
          "mtime": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mtime"
          }
        },
        "type": "object",
        "required": [
          "file_path",
          "title",
          "body"
        ],
        "title": "Note",
        "description": "Full note contents including frontmatter metadata and Markdown body."
      },
      "NoteMetadata": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "person_note",
              "decision",
              "idea",
              "observation",
              "reference",
              "meeting_note",
              "project",
              "action_item",
              "weekly_summary",
              "other"
            ],
            "title": "Type",
            "default": "other"
          },
          "template": {
            "type": "string",
            "title": "Template",
            "default": "blank"
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "default": "personal"
          },
          "people": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "People"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "source": {
            "type": "string",
            "enum": [
              "web",
              "voice",
              "teams",
              "mcp",
              "import"
            ],
            "title": "Source",
            "default": "web"
          },
          "action_items": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Action Items"
          },
          "created": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created"
          },
          "updated": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 1.0
          },
          "confidence_rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Rationale"
          },
          "review_status": {
            "type": "string",
            "enum": [
              "pending",
              "approved"
            ],
            "title": "Review Status",
            "default": "approved"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "approved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "approval_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "auto",
                  "manual"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Mode"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/LinkRef"
            },
            "type": "array",
            "title": "Links"
          },
          "org": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Org"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "NoteMetadata",
        "description": "Parsed frontmatter of a note as written/read by the ingest pipeline."
      },
      "NoteMoveRequest": {
        "properties": {
          "to_path": {
            "type": "string",
            "title": "To Path"
          }
        },
        "type": "object",
        "required": [
          "to_path"
        ],
        "title": "NoteMoveRequest"
      },
      "NotePatchRequest": {
        "properties": {
          "updates": {
            "additionalProperties": true,
            "type": "object",
            "title": "Updates"
          },
          "if_mtime": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "If Mtime"
          }
        },
        "type": "object",
        "required": [
          "updates"
        ],
        "title": "NotePatchRequest",
        "description": "Body for PATCH /api/notes/{path} \u2014 frontmatter-only patch."
      },
      "NoteRef": {
        "properties": {
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "enum": [
              "person_note",
              "decision",
              "idea",
              "observation",
              "reference",
              "meeting_note",
              "project",
              "action_item",
              "weekly_summary",
              "other"
            ],
            "title": "Type",
            "default": "other"
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "default": "personal"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "created": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created"
          },
          "updated": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 1.0
          },
          "review_status": {
            "type": "string",
            "enum": [
              "pending",
              "approved"
            ],
            "title": "Review Status",
            "default": "approved"
          }
        },
        "type": "object",
        "required": [
          "file_path",
          "title"
        ],
        "title": "NoteRef",
        "description": "Lightweight reference to a note (used in listings and search results)."
      },
      "NoteWriteRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body",
            "default": ""
          },
          "metadata": {
            "$ref": "#/components/schemas/NoteMetadata",
            "default": {
              "type": "other",
              "template": "blank",
              "domain": "personal",
              "people": [],
              "tags": [],
              "source": "web",
              "action_items": [],
              "confidence": 1.0,
              "review_status": "approved",
              "links": []
            }
          },
          "if_mtime": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "If Mtime"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "NoteWriteRequest",
        "description": "Body for PUT /api/notes/{path} \u2014 full note write."
      },
      "Page_NoteRef_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NoteRef"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 50
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "Page[NoteRef]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}